requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
billiard==4.2.3
blis==1.3.3
cachetools==6.2.2
catalogue==2.0.10
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.13
distro==1.9.0
flower==2.0.1
google-auth==2.43.0
google-genai==1.52.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.14.0
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
kombu==5.5.4
MarkupSafe==3.0.3
murmurhash==1.0.15
numpy==2.3.5
openai==2.8.1
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.2.250926
pluggy==1.6.0
portalocker==3.2.0
preshed==3.0.12
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.1
redis==7.1.0
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
tenacity==9.1.2
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tornado==6.5.2
tqdm==4.67.1
typer-slim==0.20.0
types-pytz==2025.2.0.20251108
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
websockets==15.0.1
wrapt==2.0.1
